pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rice-agents"
version = "0.1.1"
description = "A modular, provider-agnostic framework for building and orchestrating AI agents."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Shankha", email = "your.email@example.com" }]
dependencies = [
    "anthropic>=0.75.0",
    "chromadb>=0.5.0",
    "google-genai==1.56.0",
    "nest-asyncio>=1.6.0",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]

[project.urls]
Homepage = "https://github.com/yourusername/rice-agents"

[dependency-groups]
dev = ["ruff>=0.14.9", "ty>=0.0.3", "pytest>=7.4.0", "pytest-asyncio>=0.23.0"]

# --- Ty Configuration ---
[tool.ty]
python-version = "3.11"
python-platform = "darwin"

[tool.ty.src]
include = ["src", "tests", "examples"]

[tool.ty.rules]
possibly-missing-import = "warn"

# --- Ruff Configuration ---
[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src", "tests", "examples"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
]
ignore = ["E501"] # Ignore line length violations (let formatter handle it)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
